make sbrk() robust against concurrent heap_pos modification
As would occur, say, when get_kern_page() finds a page that isn't already
mapped to a kernel address, and calls reserve_heap_page() to find space for
that mapping.

This issue would've strictly been fixed by also changing get_kern_page() to
map the page at a given logical address, as that would've taken a different
path than the one that maps an unattached page to a heap address.