add proper minimal runtime to sigma0
I.e. setup a stack as part of the ELF binary. This is the simplest way;
although the upcoming spawn_sigma0() could map a couple of pages right in
front of the binary's starting position and set sp accordingly, this would
take up more code space there than it does in sigma0.