move utility modules into lib/
Slab allocation, hash computation, vsnprintf(3), and so forth. This is for the
bootloader's benefit, so the CCAN modules or dlmalloc don't end up in lib/ .

Currently dependency tracking doesn't account for image.bin and sigma0
depending on libukernel_util.a , so a full rebuild is required when library
code is changed. (use ccache.)