s0: add slab allocation, small malloc() facsimile, initial memory
The former is backed by the page tracker, which backs malloc(), which the
AVL-trees from CCAN use, which is required for the "get this definite page"
interface. Contrast with CCAN hashtables which need a proper malloc heap with
big allocations and everything; AVL trees can be serviced from a couple of
slabs.