fix another syscall wrapper bug
This time, L4_SystemClock()'s contents somehow suggest to the compiler that
%ecx stays unmodified by the code sequence. It doesn't; SystemClock returns
outputs in %eax:%edx and trashes %esi, %edi, and %ecx. Somehow "ecx" can't
appear as both an input and a trash register, though, so the solution is to
call __L4_SystemClock directly and pop an error when __pic__ is defined.