make the SystemClock syscall stub preemption safe
By detecting the case and restarting the function. The loop is guaranteed to
terminate iff there isn't a preemption event in between the load instructions
where the high 32 bits are carried into before the process context resumes.
Actual single occurrences of this on a current processor should be very very
rare, so the loop can reasonably be expected to terminate in fewer than 3
iterations.