define the return_to_*() functions as interrupt/exception exits, remove "regs" parameter
On some architectures the design that just loads the regs parameter might be
the only way to do this. But not on the x86. (The functions could be
respecified to either change *regs, or perform a nonlocal exit, for such
architectures, and the NORETURN spec can be configured in a header file.)