testbench: add test for preemption exceptions
Still not a formal one, so here's a description: when preemption works[0], the
spinner thread should be preempted a few times. On each of these times it
should send a preemption exception to the exception handler thread, which
should therefore print out a few "got exception of %u words" messages before
the "regular spinner exit" one.

[0] as far as I've understood properly