testbench: adjust sched_test.c to check an additional ThreadSwitch property
This property being that ThreadSwitch from a higher-priority thread to a
lower-priority thread essentially overrides the scheduler until such a time as
the lower-priority thread is preempted by a higher-priority thread.

In the test case this is verified, up to the first "until", by starting the
spinner thread at a lower priority than the parent thread. If ThreadSwitch
didn't override regular scheduling the call to ThreadSwitch would only donate
the parent's remaining quantum, enter the scheduler, grant the parent another
quantum, and resume the parent.

The pass/fail criteria remain unchanged. The second part, where the spinner
thread can be preempted by a higher-priority thread, should be added to the
test as well.