sched: send non-delayed preemption exceptions
This barfs and does unexpected things when delay is enabled. Aside from that,
it provokes a curious quirk of IPC timeout handling: when preemption
exceptions are enabled in sched_test, the spinner's exception handler never
receives a timeout even if its own timeout preempted the spinner. In this case
it should clearly timeout (as that is seen as having already happened),
receive the message, reply, and then wait again.