testbench: add looped tests
This is useful for things that consist of a bunch of related cases that can be
tested with the same code. There is a problem, though: nothing ensures that
the tcase_add_loop_test() call sets a sufficiently large range for the test.

(and that a number of things in sched_test() seem to completely break with
this changeset, or the immediately preceding one.)