add delayed preemptions
The unit test appears to disagree with the implementation with regard to
whether a thread calling L4_ThreadSwitch() voluntarily during a preemption
delay should experience asynchronous preemptions, or not.

The spec seems to say that ThreadSwitch should throw away the current thread's
quantum. This seems reasonable and suggests the unit test is wrong.