testbench: run fixtures in threads
This allows for failure reporting. As there's no format for communicating
fixture failure to report.pl (nor a useful policy of what to do with them),
it'll just break "make check" when that happens, which is clear enough.

Figuring out quite what to do with a failed fixture is a thing that'll wait
until report.pl learns to restart the emulator on command.