test: rewrite poke_peek_fault_test() to work with forkserv
I.e. it must allocate some fresh address space with sbrk() in order to get
proper faults. It could also use Unmap to remove that memory to ensure faults,
but as Unmap isn't implemented, this works as well -- and ends up testing just
one mechanism instead of two.

This changeset also updates the legacy thread test to add its threads to
forkserv paging, which is very necessary for thread stacks.