rework the thread API
Previously it confused the concepts of starting, stopping, resuming, going to
sleep, waking up, and IPC cancellation. No more.

Plenty of callsites were updated. A common theme is to assert that @t->state
== TS_STOPPED after calling thread_halt(t); this requires that the thread
wasn't in an IPC sleep before, which is what the old thread_stop() would break
if it was the case. The assert ensures that the old callsite was translated
correctly.

With this changeset, ExchangeRegisters should properly break, or not-break, a
waiting IPC. However the tests hit a failing assert() with regard to the
post-exception callback mechanism, which is hugely ugly and will be replaced
in about two changesets' time.