testbench: move dlmalloc.o up in the dependency list
This means it'll (start to be) compiled earlier, giving the other targets a
chance while it does.

dlmalloc.c is the largest source file in this tree and takes the longest time
to compile; this changeset represents an useful concurrency enhancement to the
tune of about a second and a half on two- and four-thread hosts.