testbench: add fork, indirect new_thread
The latter is slightly nontrivial due to the per-space privilege mechanism in
L4.X2: forkserv would like to call a helper thread in the original testbench
rootserver, which is paged by forkserv. This classic IPC dependency makes the
resulting pagefault message look like it was a reply to the dependant IPC,
which is especially funny when the next IPC call looks like a response to the
pagefault message.

This issue is solved with a helper thread and a simple queue operated with
atomic instructions. The forkserv main thread also responds to a one-way
"about to sleep" message to combine IPC wait sleep with a "happens-after"
check of per-instruction atomic state, solving the read-to-sleep race
condition.

This changeset works up until the child process wakes up. It should receive a
wakeup message from the parent process but doesn't: additionally, as its KIP
will be misconfigured a call to Ipc would segfault (albeit correctly). To make
the tests complete, the parent-side of fork() uses a 200ms time limit on the
initializing send. This should be removed once the child does pick up.

Threads created via forkserv will have priority 80 so that they don't start
before their pager TCR has been re-set right after a breath-of-life. This is
strictly the wrong thing to do, as breath-of-life should be sent in response
to a CHOPCHOP_REPLY, but w/e. Call it a TODO.

Forkserv still has a bit to go until it'll start really requiring a working
Unmap.