add first prototype of Unmap
As expected, the bulk of this changeset happens elsewhere: mapdb_unmap_fpage()
gains an "immediate" boolean argument which determines if the range given is
to be unmapped itself or not. With "recursive" this makes a four-way mode
parameter where one setting is without use for the time being. As such it's
been disallowed by assert().

mapdb_unmap_fpage() also no longer panics on recursive unmap so that the test
suites can finish properly.