mapdb: add child tracking, recursive unmap
Children are now stored in a hash table of a constant maximum probe depth,
with tombstones. It'd be possible to add a child_in_ent() query with this, but
for now this feature is only used to find (and skip) duplicate children.

mapdb_unmap_fpage() gained a parameter: clear_stored_access. Its effect has no
unit tests, and therefore it's safe to say it likely doesn't work correctly.
Callsites were updated accordingly.

deref_child() and mapdb_add_child() were added. The first dereferences a child
entry and checks it for validity; the second adds a child to the hash table.