testbench: move restarting behavior into a role
This is a role rather than a subclass due to being an amendment to Sink's
functionality; rationale being that there'll be a set of environment variable
flags for picking out the desired set of test-progress tweaks, which are
cleaner to apply to the Sink instance than to an inheritance hierarchy.

Also, a `qcheck' toplevel target was added. This omits the Mung::Restarting
role from the Mung::Sink instance via a perl-true value in the TEST_QUICK
environment variable.