mapdb: add reparent_children(), support code
This version is sadly as broken as the mapdb spec as of this changeset, i.e.
mostly, but will get better. There's also a big comment about error handling
at its call site.

For support code, add get_map_db(), which blows a BUG_ON() when the ref_id
isn't found. The check could likely just be an assert too.