mapdb: remove parent_offset from deref_child() output
That field's meaning was unclear anyhow.

Call sites were adjusted accordingly, which led to addition of two panic()
calls to mapdb.c, both indicating a FIXME: one in the distribute_children()
case where a child should be discontiguated so that it'll no longer violate
the single parent rule, and the other in the mapdb_unmap_fpage() complex case.