mapdb: ignore children out of unmap range
Previously this'd confuse the unit test by unmapping not only the page that
was expected, but also a few others in the same map_entry. Doing so led to
spurious pagefaults being logged, failing the unit test (albeit for faults on
a different page than suggested by the test itself).

This suggests the pagefault handler thread should also log the scale 12 fpage
where the fault occurred, and their access mask.