testbench: make UTCB location consistent on fork() child side
This means allocating the fork() caller's child-side thread in the same slot
as the caller, and giving it the same position in the new space's UTCB area.

forkserv was altered to accept a slot number. It's honestly not known right
now if this means anything, or if the actual UTCB location is in fact
determined by start_thread_long() .

process.c plays along with the forkserv interface change, though it could do
better by making sure the process entry thread is allocated the same
implicitly non-conflicting slot as the fork() caller's exception receiver
thread. This improves *snrk* reliability *giggle*.

process_suite.c now diag()s the UTCB locations which this changeset sets
straight. They could be tested for and against (i.e. that the
__L4_Get_UtcbAddress() non-volatileness is in fact entirely valid here), but
this requires the unpalatable shoveling subprocess test point results to the
parent process over IPC.

thread.c initializes itself by not placing either the base_tnum or utcb_base
at a two-thread offset from the one that calls init_threading(). (that one is
always the main() thread, of course.) thread_on_fork() now passes the caller's
thread number to forkserv:new_thread.