testbench: add support for forking child processes
This is a somewhat large changeset.

First, FORKSERV_EXIT now exits the whole process. It's rather untested, but
seems to pass the existing checks. A FORKSERV_EXIT_THREAD label was added, and
that does what the other used to do, but can be used to zap specified threads
rather than always killing the caller. This feature is used in join_thread(),
which now works in subprocesses and is used to catch the child_starter_fn()
thread once it's done.

Second, testbench processes can now check whether they're the original
roottask (or privileged, anyhow) by calling the is_privileged() function. This
is used in start_thread_long(), as such a process must create threads by
calling to forkserv.

Third, the child process' fork() context is now synchronized with
child_starter_fn()'s thread to make it conform to the join_thread() protocol,
i.e. x86 #GP for int $1; see $paragraph =~ /^First/ .

Fourth, start_thread_long() and thread_wrapper() now pass the parent thread ID
in the stack rather than the thread's UDH. This is necessary because the
forkserv interface doesn't permit setting it before the thread has launched;
the stack method is compatible with just "start with this sp/ip".

Fifth, assorted minor changes.