sched: add TS_XFER to IS_READY()
Also change the active receive post-exception path to not assert for
IS_READY(), as post-exception hooks may start the next IPC right away so the
"ready to schedule" transition might've'nt happened.

There's a barely-related TRACE() macro update in sched.c, which is more to
include that file in this changeset (as IS_READY() affects timeslice
expansion) than for a conventional reason.