acpi: add first ACPI table scanner
It explores those tables that're passed by the Bochs BIOS, with the exception
of the bytecode-bearing ones. For now, the goal is to scan enough to discover
the number of CPUs (though we'll only expose one via the KIP), the I/O APIC
configuration, the HPET's details, and frequency information for
ProcessorControl, KIP, and RDTSC-based timing.

These bits will be exported in <ukernel/acpi.h>, added in this changeset.

Additionally the ACPI tables should be scanned thoroughly enough to reserve
all ACPI memory for the OS personality, and mark it as special so it doesn't
get used as regular memory.