defer interrupts that happen while the CPU is in kernel code
Now, interrupts sent in response to an ISR being invoked on the kernel stack
will use deferred signaling unless received during the "hlt" instruction. This
just leaves the "exit syscall to preemption winner" mechanism, and testing.

This aggregate changeset had a long and sordid history. Hopefully wackness
such as it represented will not be seen in the future.