add return_from_exn() for latent interrupt signaling
It goes at the end of syscall and other exception handlers' C halves, and in
return_to_other() and return_to_scheduler(). It's got two effects: first, it
processes deferred interrupts (though see below); and second, it returns in a
condition where interrupts are disabled and kernel_irq_deferred is false; so
afterward, the kernel is in a state where deferred signals have been processed
and won't occur until the CPU transitions from user mode back to supervisor
mode again.

As of right now, return_from_exn() has no affordance for pre-emptions due to
interrupt signaling. It should. (int_trigger(), too.)

Interrupt handlers that don't re-enable interrupts should not call
return_from_exn(). Ones that do may disable interrupts with it (but in the
future it'll be a pre-emption path) to neatly avoid lost interrupts.