testbench: interrupt: add context test case
I.e. interrupts from different CPU contexts. int_from_wait is supposed to have
the interrupt occur in the "hlt" instruction; int_from_user, in the middle of
nsleep(); and int_under_other in the middle of another thread's nsleep()
(which should cause an immediate preemption).

Alas, it appears that a serial port is not an adequate interrupt source, at
least not under KVM -- which makes total sense given the I/O virtualization.