testbench: thread: make tid_stomp narrower
This changeset removes the test for modification of a thread ID's version bits
while referencing the address space with the ID that's being modified. That
ought to be valid, but tid_stomp isn't the right place to test it. (and I'm
fairly sure that sys_threadcontrol() doesn't do the right thing, either.)