clean up the rest of thread_find()
The one in scheduler_loop() is entirely legit in the sense that
*scheduler_mr1 can never be a local thread ID.

The redirector TID validation (i.e. that it must exist) in sys_spacecontrol()
isn't required. But there should be a check that it's not an interrupt or
kernel thread.

Finally thread_find() was flattened down to just the htable_get() call and its
parameter computations.