ipc: remove calls to thread_get_utcb(), change set_ipc_return_thread() contract
These two changes are somewhat unrelated.

The first replaces many, many redundant calls to thread_get_utcb() with
parameter passing and better structuring of the relevant code. These changes
affect only the main sys_ipc(), kipc(), etc. paths; typed transfers,
resumption, cancellation, and that lot are largely unaffected (for now).

The second moves the IS_KERNEL_THREAD() check out to the only callsite to
set_ipc_return_thread() that's affected, which is in the active success path
of ipc_send_half(). Others have an assert to the contrary, with justification,
before the call.