ipc: cleanups, move init_ipc() towards start of module
There's plenty of #if 0'd segments left in due to the incomplete nature of
string transfers, especially in the face of pagefault handling and xfer
timeouts. Those should be tested for and implemented; after that, the debugs
come out.

Also, an assert in ipc() against @utcb was changed to a NULL comparison,
because once address space recycling is implemented, getting the value of an
UTCB pointer may cause that pointer to come into existence. At that point UTCB
validity should be tested with a minimally invasive IS_UTCB_VALID() macro
instead.