x86: add recursive irq disable, tighten related asserts
This caught an invariant violation where return_from_exn() was called in
return_to_scheduler() with interrupts disabled when it was entered from the
timer interrupt's preemption clause. While that doesn't hurt, it revealed that
interrupts weren't being processed through some callers of r_t_s(), (i.e. the
extra long irq0 case), which they should. The cost of fixing this is a
somewhat messy test-and-enable before rfe().