thread: make space->utcb_pages alloc even lazier
A quick game of "hide the calloc()", so to say.

This changeset's gain is mostly in documentation: it says that
t->utcb_ptr_seg != 0 implies sp->utcb_pages != NULL, which is ensured by the
same function having run when some thread in @sp had utcb_ptr_seg == 0 in the
past.