thread: deletion idempotence, better SpaceSpecifier validation
The former means that a L4_nilthread SpaceSpecifier will always return as
though it had deleted a thread if one didn't exist.

The latter is a change from L4_Version(spacespec) == 0 to
L4_IsLocalId(spacespec), which is obviously correct.

Some trace statements were also removed, and the spacespec == dest_tid check
was moved within the block that calls space_new() .