ipc: fix non-presence detection in sender lookup
That's to say, this changeset affects ipc_recv_half(). Considering the
aggregate level of ugly in find_local_sender(), it's likely there'll be a
nicer data structure for this in <struct space> ere long.

Additionally, Ipc was changed to reject FromSpecifier in the kernel range;
this may not comply with the spec, but on the other hand there's no use case
for receiving from a particular kernel thread (and no "any kernel thread"
wildcard, either).