test: string: add tests for simple message overflow
I.e. two string items when the receiver only accepts one, and a string item
that's too long for the receiver's buffer.

Missing cases: a receiver that's got string items in its BRs but BR0 doesn't
indicate string reception; and one where the bit is set but there are no
string items. Also the same for compound strings.