test: self: add deeper_fork
The dial-a-forkbomb test. Currently it can only run with depth=5 because the
microkernel runs out of kernel memory with larger numbers, but that should
change.

A variant could be written that tests fork_tid() in this process, but it won't
terminate because of IPC sender/receiver crossing. The problem is that
fork_tid() is written in terms of fork() and not the other way around.