ipc: change abort_waiting_ipc() to accept a ThreadId
Instead of a <struct thread *>. This is because thread deletion destroys the
thread structure before IPC hooks are aborted, and because the whole thread
structure isn't actually required.

Also, ipc_user() was altered to always set the global thread ID in
ipc_{to,from}, which is necessary for matching in abort_thread_ipc() and
abort_waiting_ipc(). This convention should be followed by sys_ipc() and every
other mutator of those fields as well: ordinary IPC should also be aborted
when the peer is lost.