test: ipc: add basic Lipc tests
They work while the Lipc syscall is wired to do Ipc instead, so that's OK.

Some of the test points and variations were picked backward from a development
branch, and so are designed to demonstrate a particular bug in the
implementation. This is immaterial until someone breaks up the basic_lipc and
basic_invalid_lipc tests according to some proper criteria, but in case they
kind of look funny, you now know why.

Some room for improvement remains even within these two tests.