testbench: add a report script test & meta suite
This does an actual end-to-end test up 'til the completed Sink instance would
be passed to the reporting module. The idea is that as previous tests confirm
that Sink, Ctrl, ProcessModule, and the like work, this way testbench's parts
(the ones that generate the actual output) may be tested using bits that've
been already written and tested.