add kernel self-test support, basic test in kth.c
This changeset modifies testbench's tap.c so that it can be compiled with the
kernel's header to validate its prototypes the way the kernel sees them. The
testbench Makefile was changed to strip __KERNEL__ from CFLAGS, so that tap.c
can recognize how it's being compiled and drop _fail_unless() accordingly,
since it refers the testbench-exclusive flush_log().

When self-tests are enabled, testbench will do the invocation dance with
sigma0's pager with

    $ TESTBENCH_OPTS="ktest=1" ./run.sh -display none

as expected. No support for running the tests through the reporting script,
however, that's the very next thing.