ipc: change limit parameter to stritem_first()
It's now the number of words, rather than the maximum offset. The latter was
rather odd because the parameter was given as an unsigned integer, thereby
forbidding iterator initialization for an absence of a string item.

Similarly the "max" field in <struct stritem_iter> was changed to an int8_t .