space: unmap old KIP in space_set_kip_area()
This is a rare code path that's so hard to test that it'll just be assumed
correct.

The sole use case would have an OS personality mapping a "syscall glue" page
in somewhere, which'd be rewritten in lockstep with the relocated KIP;
presumably all threads in the space would be stopped using ExchangeRegisters,
however, any thread that was in a system call when the KIP moves will return
into non-KIP code (or just a hole in the address space). This could be avoided
by swapping out such threads' EIP bits to refer to the new KIP address, but
again, too hard to test so let's punt.