cpu: add cpu_has_sysenter(), the use_sysenter variable
It's obviously specific to x86 despite the naming.

The variable should be replaced with a static inline function that can have
two modes: constant according to compiletime kernel support (combined with
refusal to enter userspace on CPUs that haven't got SYSENTER), or dynamic (per
stored result of cpu_has_sysenter()). None of this is at all urgent however.