arch: x86: add fastpath for Ipc, Lipc to _sysenter_top
The fastpath's effect is minute at most. The benchmarks show about a 30 clock
advantage for completed closed IPC (i.e. two calls to Ipc and at least one
exit), which is about what we'd expect to see: skipping a handful of pushes
and such to construct a full exregs frame.

Now that those syscalls no longer wind up in sysenter_bottom(), it doesn't
need to have a case for them.

[This changeset was originally developed in january 2014, which is reflected in
the timestamp. Further clean-ups and updates were applied in june 2015.]