add catching of in-kernel faults, segment hack in space_memcpy_from()
This cuts in-VM KernelInterface from 40k cycles to less than 8k, i.e. a fifth
of the previous.

Once copy_stritem() is extended to spot intraspace transfers, this should
speed those up as well in the case where all affected pages are already
mapped; furthermore, the interface could be used to catch in-transfer faults
in one half of an interspace transfer without any probing.