lib: minor fixes to string.c
Notably, if string.c is compiled with -O3 in GCC 4.8.2, memset() ends up with
a call to itself. That's right: perpetual recursion because the optimizer
doesn't realize that this breaks everything. GCC 4.7 doesn't do that, so the
more recent compiler must be at fault. (not that -O3 does a lot compared to
-O2 anyway.)