sched: exit Ipc/Lipc/ThreadSwitch with SYSEXIT where available
There are two cases: switch_thread_u2u() for ThreadSwitch[0], which is
confirmed to occur three times in sched/yield_timeslice_test; and
switch_thread(), which goes from kth (e.g. the scheduler) to userspace via
sysexit where applicable[1].

The mechanism adds a TF_SYSCALL flag which is set whenever kernelspace is
entered with SYSENTER and the frame's %eip register is set to the appropriate
SYSEXIT epilogue, and cleared when the userspace context is returned to.

This changeset also includes some miscellaneous fixups to the sysenter branch,
such as removal of a redundant debug print in isr_exn_exregs_sc_bottom() and a
deduplication of the unimplemented MemoryControl syscall.

[0] and in the future, co-scheduled Ipc and Lipc
[1] i.e. when userspace didn't enter the kernel because of an interrupt