ipc: move sender state transitions out of ipc_send_half()
The problem being solved is this: there are actually three kinds of IPC that
ipc_send_half() is used for. First, IPC from kth to userspace; then IPC from
kernel to userspace for pagefaults and exceptions; and finally IPC between
userspace threads.

These three contexts handle success differently. For example, all but the
pf/exn mode go to scheduler on preemption, and the latter always returns to
the IPC peer. In this light the way ipc_send_half() always shifts the caller
into R_RECV is unnecessarily general; it also causes a bunch of control-flow
dependency in the sq_update_thread() call.

Another changeset like this will follow for ipc_recv_half(), eventually.