test: ipc: add preemption parameters to send_preempt
Now it has flags for whether preemption occurs at all (as before), whether the
victim will attempt to delay the inevitable, whether the sensitive priority is
high enough, and whether the delay is long enough for the victim's spin.

Predictably the tests fail when preemption should've been delayed, and
therefore also when the delay should've run out during the lower-priority
spin.