sched: handle R_RECV in return_from_exn()
This shouldn't be necessary if the preempting thread is waiting closed for the
current thread, because an active send will roll right over a R_RECV anyway.
In fact, that case should be caught in might_preempt().

This changeset fixes the inter-space IDL benchmark, which would otherwise see
R_RECV preemption turning into an omitted receive phase and therefore errors &
failures.