test: ipc: test for disappearing VirtualSender in passive send
It should cause propagation to fail when the active receive finally happens.

This currently doesn't cover reception by other FromSpec values than
anythread. Another test should be written to try WaitLocal, and another to
perform the propagating IPC from a different process. A further test should
receive with a definite global/local TID: after the VirtualSender goes away,
the active receive should match the propagator's TID and only that.

There are tests for correct TID form in cancel_virtualsender, which won't mean
much when the propagator and propagee are never placed in a different address
space, but perhaps that'll be a later feature.

There's a forking mode in cancel_virtualsender which should be imitated for
cancel_sender also.