testbench: accept ":*" as test iteration parameter for "runonly"
Without such a parameter it runs the first iteration. This is a change from
the previous form where the number's absence would refer to iter=0 regardless
of whether the test is even valid with that.

This suggests that the runonly syntax could accept references to testcases and
entire suites instead of shorthand for an entire test. That'd be more useful
for the test reporting script as well.