space: modify the page table in space_get_utcb_page()
When a new UTCB page must be allocated, it should be mapped into the address
space immediately.

However, it's also the case that when that page isn't mapped, the pagefault
handler should perform the exactly equivalent operation through mapdb. This
causes certain tests to fail, and the indirect ThreadControl operation in
forkserv's end_thread() to scream about failures and such. That symptom, as
well as the "many faults" report from forkserv in the UTCB range, suggests a
fuckup with regard to either the entire UTCB page, just the "UTCB area
pointer", or even the %gs register.