testbench: make kill_thread() idempotent between calls to start_thread_long()
That's to say: calling it multiple times on the same TID, or after a
xjoin_thread() on the same TID, won't cause immense breakage. However, a FIXME
was added for the manager thread's curious behaviour. It should be considered
in detail.

Also, report invalid (out-of-space) parameter to kill_thread(). This makes it
slightly more informative before the assert blows us out of the sky.