test: ipc: add receive_from_local
Very basic interface stuff. The goal is to have every nook and cranny of the
Ipc syscall covered so that those tests can be modified to instead call Lipc
in another test case, or variant suite (if that mechanism turns out to be sane
enough).

Next up: the sibling test, receive_from_foreign.