testbench: add small application benchmark
However, as this is a microkernel project, the application is an operating
system primitive. fork/wait in fact.

From the benchmark, it seems to scale linearly with respect to the number of
children. That's good: this means that there's no idling. It's also bad
because a single fork/wait pair takes 1.3M cycles on Nehalem KVM. Eager
filling of page tables should bring that number down.