mapdb, space: small range predicate tweaks
ADDR_IN_FPAGE() was changed to use fpage_overlap() between the given fpage and
an invalid fpage which is nevertheless accepted by fpage_overlap(). This lets
GCC inline the MIN() clause away in the latter.

With that gain in mind, a BETWEEN() was replaced with an ADDR_IN_FPAGE() in
deref_child().

sys_spacecontrol() now tests overlap between kip_area and utcb_area with
fpage_overlap() for clarity, and cares about @utcb_area for correctness.