arch: add uniform page table access interface, i386 impl, use case
The use case is a postcondition for mapdb_add_map() that checks that the
affected page table entries are either not-present or set to the range of page
IDs given to the function.

arch_x86.c was added. Things that're namechecked in <ukernel/x86.h> should be
moved into it. Eventually it'll be renamed to conform to fresh new
directory-structure ideals, but until then a prefix is fine too.