space: fix a null-pointer niggle in sys_unmap()
This was exposed by the clang static analyzer. Despite this being in debug
code and thus seemingly unimportant, all code is supposed not to use pgid=0 as
a special value anyway, so in it goes.

The changeset also adds a note about how alloc_tss() is horrible and should be
fixed by using valloc() instead and relying on dlmalloc's ability to reuse the
leftovers smartly.