test: redir: add redirection suite
The test suite is a bit of a mess: allocation is all over the place and the
return-message darray carries pointers for talloc. That's no good, but it'll
work.

As of this changeset `fork_redirect_deep' isn't very useful: the redirection
deadlock case should have a proper dedicated test for it which doesn't cause a
two-layer fork to lock up and leave children behind in uncertain states.
This'll get its proper development once redirection chains have been fixed.