ipc: add redir_wait sleeping on missing or unready redirection chain
This changeset has three parts in the IPC state machine: the first is in
ipc_send_half() which now checks that the sender's redirection chain is
present and ready, and goes into redir_wait if not. This happens when active
send is confirmed possible, moving the redirection checks into the active
half.

The second part alters the receive phase to match among redirected passive
senders only those where the redirection chain is ready. Those that don't are
moved into redir_wait from sendwait_hash, and one that does has its send-phase
re-executed to kick redirection off. If this causes a thread to become
runnable, the thread remains in R_RECV and leaves the rest to the scheduler.
[Notably the scheduler may pick the receiver as next thread to run and thus
cause another redirection chain to activate. This is entirely benign and will
not cause malfunction.]

The third part picks up redir_wait entries as a form of active receive. Its
helper function find_redir_sender() should rewrite the redir_wait entry when
some other thread is holding the chain up, though this isn't met by current
testing.

Another section of the changeset deals with halting in-progress IPC when a
redirector becomes invalid, and resuming execution when the invalid state is
exited either by reinstating a redirector, or clearing the redirection
setting. Resumption causes re-evaluation of the redirection chain through
re-execution of ipc_send_half(); this interface is also used in active receive
from a redirected sender.

This changeset adds three data structures: the `redir_wait' htable in ipc.c,
the TF_REDIR_WAIT flag in <struct thread>.flags, and <struct
thread>.u1.waited_redir as parameter for the REDIR_WAIT mode. The flag
specializes the SEND_WAIT scheduling state to re-use its timeout behaviour.

As of this changeset, basic_redir has some failing test points. These have to
do with the way that the sender behaves when the redirector holds a message
back; the sender should only move into its receive phase (or exit IPC) on
timeout or when a message is propagated to the intended destination on the
sender's behalf. This preserves transparency in error reporting.