test: redir: add changing_redirector
Also move get_redir_msgs() out of another test.

This test is presented before the actual change because it would otherwise
trigger a panic in the kernel, preventing it from running properly. The fix
corrects this.

There's an useless case, !active_send && redir_wait, which is left unskipped
for validation's sake. It may be that a future development series skips over
it.