space: make redirector-switching consistent in SpaceControl
This works by first invalidating a previous redirector, and then re-starting
threads once the new redirector is in place. That causes extra work (O(n) to
the number of redir-waiters) which may be significant to use-cases where
redirector threads are being balanced in response to scheduling, or something
like that. So it's not important for v0 at all.

Notably this solution doesn't work for redir_waits from spaces that're
redirected by a thread in the primary space, so chains with more than one
redirector will still lead to breakage.